Move getOrientedDimensions and calculateCenteredCrop to their own file
