+++ src/utils/adjustments.tsx
@@ -691,42 +691,3 @@ export const ADJUSTMENT_SECTIONS: Sections = {
     Effect.VignetteRoundness,
   ],
 };
-
-export function getOrientedDimensions(
-  imageWidth: number,
-  imageHeight: number,
-  orientationSteps: number,
-): { width: number; height: number } {
-  const isSwapped = orientationSteps === 1 || orientationSteps === 3;
-  return {
-    width: isSwapped ? imageHeight : imageWidth,
-    height: isSwapped ? imageWidth : imageHeight,
-  };
-}
-
-export function calculateCenteredCrop(
-  imageWidth: number,
-  imageHeight: number,
-  orientationSteps: number,
-  aspectRatio: number | null,
-  rotation: number = 0,
-): Crop | null {
-  if (!aspectRatio || aspectRatio <= 0) return null;
-
-  const { width: W, height: H } = getOrientedDimensions(imageWidth, imageHeight, orientationSteps);
-
-  const angle = Math.abs(rotation);
-  const rad = ((angle % 180) * Math.PI) / 180;
-  const sin = Math.sin(rad);
-  const cos = Math.cos(rad);
-
-  const h_c = Math.min(H / (aspectRatio * sin + cos), W / (aspectRatio * cos + sin));
-  const w_c = aspectRatio * h_c;
-
-  return {
-    x: Math.round((W - w_c) / 2),
-    y: Math.round((H - h_c) / 2),
-    width: Math.round(w_c),
-    height: Math.round(h_c),
-  };
-}
